requirements.txt
fastapi
psycopg2
nltk
pymorphy2
python-dotenv